Migration files only create data if the models exist in the application layers.

+++ db/migrate/20140220224204_add_hospital_to_place_type.rb
@@ -1,11 +1,15 @@
 class AddHospitalToPlaceType < ActiveRecord::Migration
   def up
-    PlaceType.create!(name: 'Hospital')
-    PlaceType.create!(name: 'Clinique')
+    if defined? PlaceType
+      PlaceType.create!(name: 'Hospital')
+      PlaceType.create!(name: 'Clinique')
+    end
   end
 
   def down
-    PlaceType.where(name: 'Hospital').destroy_all
-    PlaceType.where(name: 'Clinique').destroy_all
+    if defined? PlaceType
+      PlaceType.where(name: 'Hospital').destroy_all
+      PlaceType.where(name: 'Clinique').destroy_all
+    end
   end
 end
